Move LoggedFStream to its own file
